finished followers

+++ src/main/java/com/app/woofer/controller/UserController.java
@@ -73,6 +73,9 @@ public class UserController {
     //DELETE requests
     @DeleteMapping("/delete/{id}")
     public void removeUser(@PathVariable int id) {
+        if(id <= 0){
+            throw new WooferException("no negative id in database");
+        }
         userService.removeUser(id);
     }
 }
